requirements.txt
spacy>=3.0.0
nltk>=3.6.0
PyPDF2>=3.0.0
python-docx>=0.8.11
scikit-learn>=0.24.0
PyQt6>=6.4.0 